Fix critical compilation errors and warnings
- Add missing #include <vector> to GameState.h
- Fix MainMenu::createMenuButton signature to include x parameter
- Add UILabel::getColor() method
- Remove deprecated glColor4f call in UIRenderer
- Fix C4458 parameter hiding warnings in Engine.h, Framebuffer.cpp, Texture.cpp, BatchRenderer.cpp, UIRenderer.h
- Fix C4100 unreferenced parameter warnings across multiple files
- Fix C4244 int to float conversion in MainMenu.cpp
- Fix C4996 localtime deprecated warning in Logger.cpp

+++ cpp/src/ui/MainMenu.cpp
@@ -39,8 +39,8 @@ void MainMenu::createUI() {
     backgroundPanel->setColor(glm::vec4(0.05f, 0.05f, 0.1f, 0.95f));
     
     // Create title label
-    float titleX = screenWidth / 2 - 200;
-    float titleY = 100;
+    float titleX = static_cast<float>(screenWidth / 2 - 200);
+    float titleY = 100.0f;
     titleLabel = std::make_unique<UILabel>(titleX, titleY, "THE DAILY GRIND");
     titleLabel->setScale(2.0f);
     titleLabel->setColor(glm::vec3(1.0f, 0.9f, 0.7f));
